requirements.txt
pandas>=2.2.0,<3.0
SQLAlchemy>=2.0.25,<3.0
psycopg2-binary>=2.9.9,<3.0
python-dotenv>=1.0.1,<2.0

reportlab>=4.1.0,<5.0

openai>=1.40.0,<2.0
